Fix sizeToFit for JMTabView
JMTabContainer's sizeThatFits: method was dependent on a side-effect of layoutSubviews, so it would never actually be useful until after it was drawn.

If sizeThatFits would yeild a zero size, double check and make sure it's up to date before returning the size of the contained objects.

After that, simply delegating JMTabView's sizeThatFits: to the JMTabContainer allows the JMTabView to be properly resized to fit its subviews.